+++ pubspec.yaml
@@ -42,11 +42,15 @@ dependencies:
     sdk: flutter
   intl: any
   google_fonts: ^6.0.0
+  provider: ^6.1.1
 
 
 dev_dependencies:
+
+
   flutter_test:
     sdk: flutter
+  mockito: ^5.4.0
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
