+++ test/theme_mode.test.dart
@@ -1,0 +1,14 @@
+import 'package:chat_app/main.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  test('isDarkMode toggles value', () {
+    expect(isDarkMode.value, false);
+
+    isDarkMode.value = true;
+    expect(isDarkMode.value, true);
+
+    isDarkMode.value = false;
+    expect(isDarkMode.value, false);
+  });
+}
